Add gas env
Fix gas split

+++ functions/src/utils.ts
@@ -4,7 +4,7 @@ export const RPC = process.env.SUI_RPC ?? 'https://mainnet-rpc.releap.xyz:443'
 export const TX_WINDOW = 500
 
 export const GAS_COUNT = parseInt(process.env.GAS_COUNT ?? '20')
-export const GAS_AMOUNT = parseInt(process.env.GAS_AMOUNT ?? '1')
+export const GAS_AMOUNT = parseFloat(process.env.GAS_AMOUNT ?? '1')
 
 export async function getAllOwnedObjects(provider: JsonRpcProvider, address: string) {
     const data: PaginatedObjectsResponse['data'] = []
